Improve SVG management
Two benefits: actually release the pixbuf so we don't leak memory
if updateUI is called an excessive amount of times, and load the
filled-in template directly from RAM, without needing to write
an auxiliary SVG to disk.

Actually from like yesterday, but I wanted to rebase a fix for the old
version, which I didn't have the heart to leave broken.