Learn that calling unref in gjs is a BAD IDEA
It logs a CRITICAL in .xsession-errors in Cinnamon 6,
and it outright crashes Cinnamon 5.

I left Cinnamon idle for one hour, calling updateUI every second,
and saw no blantant memory leak or anything, so I guess it's okay.

I could have removed this with the rebase, but I wanted to leave
this in the commit history, as a lesson learned of sorts.