Remove old import system and its hardcoded UUIDs
This will make the desklet incompatible w/ older versions of Cinnamon
(5 is fine), but it'll be much easier to test and maintain.

Left the hardcoded UUIDs for gettext b/c I don't know of an easy way to
retrieve it from Python widget code, and they appear to be harmless.